# perf: add lock to client