### 概要
clientにmutexを導入しました。
primaryフラグの導入により問題になっていたデータ不整合は解消されておりましたが、provider側にロックを入れることで処理時間が短縮されました。

ロック導入前
```bash
edge_value.demo_string: Creating...
edge_value.demo_json: Creating...
edge_value.demo_bool: Creating...
edge_value.demo_json: Creation complete after 1s [id=demo-json-value]
edge_value.demo_string: Creation complete after 2s [id=demo-string-value]
edge_value.demo_bool: Creation complete after 5s [id=demo-bool-value]

Apply complete! Resources: 3 added, 0 changed, 0 destroyed.
```


ロック導入後
```bash
edge_value.demo_string: Creating...
edge_value.demo_json: Creating...
edge_value.demo_bool: Creating...
edge_value.demo_bool: Creation complete after 0s [id=demo-bool-value]
edge_value.demo_json: Creation complete after 0s [id=demo-json-value]
edge_value.demo_string: Creation complete after 1s [id=demo-string-value]

Apply complete! Resources: 3 added, 0 changed, 0 destroyed.
```